Use new-style proxied values
The dollar-sign style would be resolved by the shell which meant the order
that environment variables were defined mattered. If the shell interpreted it
first, the value would be resolve to '' (undefined) causing the value to not
be proxied correctly.

The solution is to use a shell-safe marker for proxied-values. Since Python's
using '{blah}' for new-style string formatting, this seems like a reasonable
choice.

The only other choice considered was '{{ blah }}' to match Jinja, but it's
slightly more typing, and not I think I'd rather match Python here instead of
Jinja.

The dollar-sign style of proxying is deprecated and will be removed in a 1.0
release.