Extend read_envfile to allow env vars to be overwritten
This is useful for applications that want to update their existing
env vars. If you use a system like consul-template to provide
environment variables, then it's useful to be able to update your
application when they change.

This change adds a _overwrite param to the read_envfile method,
which allows existing env vars to be updated.