Allow env file to override environment
Prior to this change, the environment always took precedence when there
was a value specified in both the env file and the environment.  This
was not suitable to all use-cases (such as those in Kubernetes, where
environment variables are commonly much less controlled and, so, much
more vulnerable to malicious value injection).

This change allows the developer to choose which takes precedence, the
environment or the env file (with the current behaviour remaining the
default).